backend/requirements.txt
Flask-Cors==4.0.0
PyJWT==2.8.0
docopt==0.6.2
flask-sqlalchemy==3.1.1
nbconvert==7.16.3
pip-autoremove==0.10.0
